Skip RGS on regtest
