+++ mutiny-core/src/lib.rs
@@ -155,7 +155,7 @@ mod tests {
             "".to_string(),
             None,
             None,
-            Some(Network::Testnet),
+            Some(Network::Regtest),
             None,
             None,
             None,
@@ -176,7 +176,7 @@ mod tests {
             "".to_string(),
             None,
             None,
-            Some(Network::Testnet),
+            Some(Network::Regtest),
             None,
             None,
             None,
@@ -203,7 +203,7 @@ mod tests {
             "".to_string(),
             None,
             None,
-            Some(Network::Testnet),
+            Some(Network::Regtest),
             None,
             None,
             None,
