+++ mutiny-wasm/src/lib.rs
@@ -826,7 +826,7 @@ mod tests {
             "password".to_string(),
             None,
             None,
-            Some("testnet".to_owned()),
+            Some("regtest".to_owned()),
             None,
             None,
             None,
@@ -850,7 +850,7 @@ mod tests {
             "password".to_string(),
             Some(seed.to_string()),
             None,
-            Some("testnet".to_owned()),
+            Some("regtest".to_owned()),
             None,
             None,
             None,
@@ -876,7 +876,7 @@ mod tests {
             "password".to_string(),
             Some(seed.to_string()),
             None,
-            Some("testnet".to_owned()),
+            Some("regtest".to_owned()),
             None,
             None,
             None,
